Update customize-webviews.md
Created a distinction between "Web Browser" and "Web Components" and removed Technology which used to be used to describe the grouping of ASWebAuthenticationSession, SFAuthenticationSession, SFSafariViewController and WKWebView. Web Component is now the only phrase used to describe this grouping. Also added a note about WebviewParameters.presentationStyle not applying to all Web Components. Note that System Web Browser implies System Web Component, but not the other way around.